requirements.txt
langchain
wikipedia
langchain-community
langchain-openai
langchain-anthropic
python-dotenv
pydantic>=2.0.0
duckduckgo-search
openai>=1.0.0